Replaced panics with an error
Go programs should not panic unless an unexpected issue arrises.
Typically they are reserved for bugs. Each call to panic in
config.go happens for a scenario that is entirely expected.

In these cases, the caller should be allowed to handle these
errors gracefully.

This change causes all such errors to be returned as an error
object, rather than panicking.