Update README.md
The README for POST/PUT is misleading. POST example does not have a body, and PUT shows a marshalled bytes for `client.NewJSONRequest` input.  This did not work for `create stream` or `create origin` APIs; the server returned 400 "Unable to read the data. Please verify the input data."
The API requires an object on input which is then marshalled at https://github.com/ap13istream/AkamaiOPEN-edgegrid-golang/blob/master/client-v1/client.go#L74
This is to add an appropriate example.